# swagger/openapi-gateway.yaml
openapi: 3.0.3
info:
  title: EnglishForU - API Gateway
  description: Central entry point for all EnglishForU microservices (port 8090)
  version: 1.0.0
servers:
  - url: http://localhost:8090
    description: API Gateway
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
...
paths:
...
  /api/packages:
    get:
      tags: [Packages]
      summary: Get all packages (ADMIN)
      responses:
        '200': { description: All packages }
    post:
      tags: [Packages]
      summary: Create package (ADMIN)
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '201': { description: Created }
...
  /api/payments:
    post:
      tags: [Payments]
      summary: Create payment
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Payment created }
    get:
      tags: [Payments]
      summary: List all payments (ADMIN)
      responses:
        '200': { description: Payment list }
...
  /api/promos:
    get:
      tags: [Promos]
      summary: Get all promo codes (ADMIN)
      responses:
        '200': { description: Promo list }
    post:
      tags: [Promos]
      summary: Create promo code (ADMIN)
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '201': { description: Created }
...
  /api/clubs:
    get:
      tags: [Clubs]
      summary: Get all clubs
      parameters:
        - name: userId
          in: query
          schema: { type: string }
      responses:
        '200': { description: Club list }
    post:
      tags: [Clubs]
      summary: Create a club
      parameters:
        - name: userId
          in: query
          required: true
          schema: { type: string }
        - name: userName
          in: query
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Club created }
...
  /api/events:
    get:
      tags: [Events]
      summary: Get all events
      parameters:
        - name: userId
          in: query
          schema: { type: string }
      responses:
        '200': { description: Event list }
    post:
      tags: [Events]
      summary: Create event
      parameters:
        - name: userId
          in: query
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Event created }
...
  /api/evaluations:
    get:
      tags: [Evaluations]
      summary: Get evaluations (ADMIN/TUTOR)
      parameters:
        - name: studentId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Evaluation list }
    post:
      tags: [Evaluations]
      summary: Create evaluation (ADMIN/TUTOR)
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Created }
...
  /api/courses:
    get:
      tags: [Courses]
      summary: Get all courses
      responses:
        '200': { description: Course list }
    post:
      tags: [Courses]
      summary: Create course (ADMIN)
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Created }
...